+++ libs/neon/src/ne_request.c
@@ -286,7 +286,7 @@ static ssize_t body_string_send(void *userdata, char *buffer, size_t count)
 	req->body.buf.remain -= count;
     }
 
-    return count;
+    return (int)count;
 }    
 
 static ssize_t body_fd_send(void *userdata, char *buffer, size_t count)
@@ -305,7 +305,7 @@ static ssize_t body_fd_send(void *userdata, char *buffer, size_t count)
         if ((ne_off_t)count > req->body.file.remain)
             count = (size_t)req->body.file.remain;
         
-        ret = read(req->body.file.fd, buffer, count);
+        ret = read(req->body.file.fd, buffer, (unsigned int)count);
         if (ret > 0) {
             req->body.file.remain -= ret;
             return ret;
@@ -922,7 +922,7 @@ ssize_t ne_read_response_block(ne_request *req, char *buffer, size_t buflen)
         }
     }
     
-    return readlen;
+    return (int)readlen;
 }
 
 /* Build the request string, returning the buffer. */
